unindent misleadingly indented line
the actual code of the loop is not dependent on this line,
and the reverse is not true either.

so if the line was supposed to be inside the loop, it would
potentially executed many times, while only the first time would
change anything.

this commit unindents the code, which retains behaviour, but
makes it clear that the code is not inside the loop.

if the loop is not entered, it could be that the previous
intention was not to execute this line. but for that to happen
n_maps would have to be 0, and what is the point of having an
'empty' texture map?

the nxt becomes negative in that case, which is probably bad.
but this commit just retains that behaviour.